tty: make tty_operations::write()'s count size_t
JIRA: https://issues.redhat.com/browse/RHEL-24205
Conflicts: tty3270 changes had to be applied by hand due
           to driver being out of date in cs9.

commit 95713967ba52389f7cea75704d0cf048080ec218
Author: Jiri Slaby (SUSE) <jirislaby@kernel.org>
Date:   Thu Aug 10 11:15:03 2023 +0200

    tty: make tty_operations::write()'s count size_t

    Unify with the rest of the code. Use size_t for counts and ssize_t for
    retval.

    Link: https://lore.kernel.org/r/20230810091510.13006-30-jirislaby@kernel.org

+++ drivers/s390/char/con3215.c
@@ -939,7 +939,8 @@ static unsigned int tty3215_write_room(struct tty_struct *tty)
 /*
  * String write routine for 3215 ttys
  */
-static int tty3215_write(struct tty_struct *tty, const u8 *buf, int count)
+static ssize_t tty3215_write(struct tty_struct *tty, const u8 *buf,
+			     size_t count)
 {
 	struct raw3215_info *raw = tty->driver_data;
 	int i, written;
